requirements.txt
flask
pdfplumber
google-generativeai
gunicorn